Split doctr into explicit 'doctr local' and 'doctr travis' subcommands
It requires --force if travis (local) is run and it detects that it is (not)
on Travis.

I'm undecided if I actually want this behavior or not. The behavior before
this detects it automatically, allowing the user to switch the behavior if
so-desired. This way is more explicit.

In any case, it's worth noting that running the wrong command in the wrong
place almost never is what you want (except for us developers doing testing).
The local command requires user input. The Travis command requires the
encrypted GH_TOKEN variable, and also makes some assumptions that are always
true on Travis (like that it is on Linux, that it will be run from the git
repo directory, and that the git remote is https).