change means of locating sphinx build directory
This should hopefully resolve #69

If `doctr deploy` is run for the first time on a repository, the old
method works fine -- it locates `conf.py` and then goes from there.
However, if it isn't the first time, then an orphan `gh-pages` branch
has already been created and the `conf.py` file has been deleted in this
branch, so the location glob will fail.

This instead looks for `*build/html` recursively, since we are assuming
the build directory to be named either `_build` or `build`