[Prototype] Allow users to deploy on their own fork/PR
This is only the first half (and still draft), instead of commiting the
encryption key, it base64 encode it and set it on travis-ci using the
API.

The second part would be to detect that the env variable is set and
trigger the right behavior, which imho is to deploy the current branch
on `<user-fork>:gh-pages` branch, under a `docs-$BRANCHNAME` subfolder.
WE can't do `$PRNUMBER` as branch will be deployed from the users run of
travis on not from the PR one.