Don't capture the command output when there is no token
This changes the API of run_command_hiding_token to only return the
returncode, and always print the output.

This change is done so that the command output in Travis appears in order.
Unfortunately, I don't know how to do this properly when there is a token, but
that's not a common use-case.