Make --command run in the deploy branch again
This reverts a regression in the 1.6 release, which made --command run from
the main branch. This is less useful, since you can just run a command before
running doctr if you want to run something in the deploy branch. Running in
the deploy branch is what's needed (e.g., SymPy's docs run a script in the
docs repo that modifies the html files before committing).

I changed the test in .travis.yml to run in the branch build, which adds a
file to the branch deploy directory (since adding a file to the index from
--command to deploy it is supposed to a supported way to sync extra files).