Clean up the sidebar table of contents
Now "Table of Contents" only appears once on the index page.

Apparently it is impossible to include the index page in the table of contents
and have it expand its subheadings in the globaltoc. You can use "self", but
it doesn't expand the subheadings. If you include "index" it complains about
recursive toc.

I had to do this manually by modifying the sidebar templates. The result is
not ideal (the index page is sidebar styled a little differently from the
rest), but it is better than before. I also added "self" so that the main page
appears on the global toc for every other page. This makes the main page
appear twice on itself. I don't know how to fix that.