Add bold black (gray) color in doctr configure
This should be used for text that should be replaced by the user before
running a command or pasting it in .travis.yml.

Many of these could perhaps be eliminated by asking about them in doctr
configure. We should especially do this if we ever get automatic .travis.yml
editing working.