Remove the note parameter to generate_ssh_key()
It was not being used by the cryptography generation. I could probably still
add it in, but as far as I can tell, the note is not visible from GitHub
anyway (and the public ssh key is not used anywhere else). GitHub shows its
own note, which is set by the title parameter of upload_GitHub_deploy_key().