Fix private repos
Factor out the code to generate a temporary GitHub token into a function. This
token is needed to authenticate with Travis to get Travis's token, and is
immediately deleted.

Also fix some other minor issues

- Remove a debug print statement
- Check for forbidden status when testing travis-ci.org

I have not yest tested if this actually works.

Also todo is to reuse the token when using --token instead of generating a
temporary one.