Encrypt the private SSH key as a Travis secure environment variable
This is to demonstrate that this doesn't actually work. See
https://github.com/drdoctr/doctr/pull/349#issuecomment-511216861. The private
key is too large to be encrypted with the Travis public key.

Instead we have to do what we are currently doing. Create a symmetric
encryption key to encrypt the actual SSH key, and then encrypt that symmetric
key into the secure environment variable (and save the encrypted SSH private
key to a file).