Double check if files still exist before git-add
If the --command flag is used, the given command runs after syncing,
but before committing. If the command removes any files that were
"added" by the syncing, or re-creates files that were "removed" during
the syncing, then the "git add" and "git rm" commands will fail.

To get around this, we can just double check if "added" files are still
there before calling "git add", respectively "removed" files are still
deleted. If there are any changes to the status after the sync, a
warning seems appropriate.