Fix user and repo tests for Travis
Not sure of the exact date, but Travis is no longer case-sensitive when
dealing with user and repo names, at least as of the date of this
commit.

Since these are the only tests in `local` we have for checking if a repo
exists in Travis, I've added a random hex-string to the end the repo and
user, respectively, to ensure that the `RuntimeError` is raised as
expected.